No more negative timings
Now that there are no more divisions by zero, the parser doesn't
instantiate negative values.

+++ timestamp.h
@@ -5,13 +5,13 @@
 
 class Timestamp{
   public:
-    Timestamp(unsigned int time, int tick, int nbpm);
+    Timestamp(long int time, int tick, int nbpm);
     int getTime();
     int getTick();
     float getBPM();
     int getNbpm(); // Weird chart formatting (133000 = 133 bpm)
   private:
-    const unsigned int time; // Time in the song in milliseconds
+    const long int time; // Time in the song in milliseconds
     const int tick; // Current tick in the song
     const int nbpm; // New BPM to change to
 };
